+++ test/hirefire/macro/test_queue_classic.rb
@@ -32,6 +32,30 @@ class HireFire::Macro::QCTest < Minitest::Test
     assert_in_delta 60, HireFire::Macro::QC.job_queue_latency(:mailer), LATENCY_DELTA
   end
 
+  def test_job_queue_latency_excludes_locked_jobs
+    queue = QC::Queue.new("default")
+    queue.enqueue_at(1.minute.ago.to_i, "BasicJob.perform")
+    locked = queue.lock
+    refute_nil locked
+    assert_equal 1, queue.count_ready
+
+    assert_equal 0, HireFire::Macro::QC.job_queue_latency
+    assert_equal 0, HireFire::Macro::QC.job_queue_latency(:default)
+  end
+
+  def test_job_queue_latency_ignores_locked_when_unlocked_due_exists
+    default = QC::Queue.new("default")
+    default.enqueue_at(3.minutes.ago.to_i, "BasicJob.perform")
+    locked = default.lock
+    refute_nil locked
+    assert_equal 1, default.count_ready
+    QC::Queue.new("mailer").enqueue_at(1.minute.ago.to_i, "BasicJob.perform")
+
+    assert_in_delta 60, HireFire::Macro::QC.job_queue_latency, LATENCY_DELTA
+    assert_equal 0, HireFire::Macro::QC.job_queue_latency(:default)
+    assert_in_delta 60, HireFire::Macro::QC.job_queue_latency(:mailer), LATENCY_DELTA
+  end
+
   def test_job_queue_size_without_jobs
     assert_equal 0, HireFire::Macro::QC.job_queue_size(:default)
   end
@@ -50,6 +74,34 @@ class HireFire::Macro::QCTest < Minitest::Test
     assert_equal 1, HireFire::Macro::QC.job_queue_size
   end
 
+  def test_job_queue_size_excludes_locked_jobs
+    queue = QC::Queue.new("default")
+    queue.enqueue("BasicJob.perform")
+    locked = queue.lock
+    refute_nil locked
+    assert_equal 1, queue.count_ready
+
+    assert_equal 0, HireFire::Macro::QC.job_queue_size
+    assert_equal 0, HireFire::Macro::QC.job_queue_size(:default)
+  end
+
+  def test_job_queue_size_counts_due_unlocked_only_with_locked_and_future_present
+    QC::Queue.new("default").enqueue("BasicJob.perform")
+    QC::Queue.new("mailer").enqueue("BasicJob.perform")
+    QC::Queue.new("default").enqueue_at(1.minute.from_now.to_i, "BasicJob.perform")
+    other = QC::Queue.new("other")
+    other.enqueue("BasicJob.perform")
+    locked = other.lock
+    refute_nil locked
+    assert_equal 1, other.count_ready
+
+    assert_equal 2, HireFire::Macro::QC.job_queue_size
+    assert_equal 1, HireFire::Macro::QC.job_queue_size(:default)
+    assert_equal 1, HireFire::Macro::QC.job_queue_size(:mailer)
+    assert_equal 0, HireFire::Macro::QC.job_queue_size(:other)
+    assert_equal 2, HireFire::Macro::QC.job_queue_size(:default, :mailer)
+  end
+
   def test_job_queue_size_with_comma_in_queue_name
     QC::Queue.new("a,b").enqueue("BasicJob.perform")
     assert_equal 1, HireFire::Macro::QC.job_queue_size(:"a,b")
@@ -61,7 +113,19 @@ class HireFire::Macro::QCTest < Minitest::Test
   end
 
   def test_deprecated_queue_method
-    QC::Queue.new("default").enqueue_at(1.minute.from_now.to_i, "BasicJob.perform")
+    QC::Queue.new("default").enqueue("BasicJob.perform")
+    assert_equal 1, HireFire::Macro::QC.queue
+  end
+
+  def test_deprecated_queue_method_excludes_locked_and_future
+    queue = QC::Queue.new("default")
+    queue.enqueue("BasicJob.perform")
+    queue.enqueue_at(1.minute.from_now.to_i, "BasicJob.perform")
+    queue.enqueue("BasicJob.perform")
+    locked = queue.lock
+    refute_nil locked
+    assert_equal 2, queue.count_ready
+
     assert_equal 1, HireFire::Macro::QC.queue
   end